Trait ferroboy::operations::operation::Operation

pub trait Operation: Sync + Debug + Disassemble {
    fn act(&self, state: &mut State) -> Result<()>;
}

An action that can be taken on the system.

Operations should, ideally, encapsulate all the changes to the system that occur as the result of an instruction. This includes things like mutating registers, changing flags, writing to memory, etc.

Required methods

fn act(&self, state: &mut State) -> Result<()>

Loading content...